* Cleanup writer test
